chore: update main layout to [lang] escope
