+++ apps/web/src/app/[lang]/layout.tsx
@@ -1,6 +1,40 @@
+import { VercelToolbar } from '@vercel/toolbar/next';
+import { NextSSRPlugin } from '@uploadthing/react/next-ssr-plugin';
+import { extractRouterConfig } from 'uploadthing/server';
+import { cookies } from 'next/headers';
+import { Lato, Source_Sans_3 } from 'next/font/google';
+import clsx from 'clsx';
+
+import { Footer, Html, ThemeProvider } from '@hypha-platform/ui/server';
+import { AuthProvider } from '@hypha-platform/authentication';
+import { useAuthentication } from '@hypha-platform/authentication';
+import { ConnectedButtonProfile } from '@hypha-platform/epics';
+import { EvmProvider } from '@hypha-platform/evm';
+import { useMe } from '@hypha-platform/core/client';
+import { fileRouter } from '@hypha-platform/core/server';
+import { HYPHA_LOCALE } from '@hypha-platform/cookie';
+import { i18nConfig } from '@hypha-platform/i18n';
+import { MenuTop } from '@hypha-platform/ui';
+
+import '@hypha-platform/ui-utils/global.css';
+
+const lato = Lato({
+  subsets: ['latin'],
+  display: 'swap',
+  weight: ['900', '700', '400', '300'],
+  variable: '--font-heading',
+});
+
+const sourceSans = Source_Sans_3({
+  subsets: ['latin'],
+  display: 'swap',
+  weight: ['900', '700', '400', '300'],
+  variable: '--font-body',
+});
+
 export const metadata = {
-  title: 'Welcome to web',
-  description: 'Generated by create-nx-workspace',
+  title: 'Hypha',
+  description: "Hypha's complete DAO toolkit helps individuals, projects, and organizations achieve more together",
 };
 
 export default async function RootLayout({
@@ -8,9 +42,53 @@ export default async function RootLayout({
 }: {
   children: React.ReactNode;
 }) {
+  const shouldInjectToolbar = process.env.NODE_ENV === 'development';
+  const cookieStore = await cookies();
+  const lang = cookieStore.get(HYPHA_LOCALE)?.value || i18nConfig.defaultLocale;
+
   return (
-    <div className="pt-9 h-full flex justify-normal">
-      <div className="w-full h-full">{children}</div>
-    </div>
+    <Html className={clsx(lato.variable, sourceSans.variable)}>
+      <AuthProvider
+        config={{
+          appId: process.env.NEXT_PUBLIC_PRIVY_APP_ID!,
+        }}
+      >
+        <ThemeProvider
+          attribute="class"
+          defaultTheme="system"
+          enableSystem
+          disableTransitionOnChange
+        >
+          <EvmProvider>
+            <MenuTop logoHref={`/${lang}/network`}>
+              <ConnectedButtonProfile
+                useAuthentication={useAuthentication}
+                useMe={useMe}
+                newUserRedirectPath="/profile/signup"
+                baseRedirectPath="/my-spaces"
+                navItems={[
+                  {
+                    label: 'Network',
+                    href: `/${lang}/network`,
+                  },
+                  {
+                    label: 'My Spaces',
+                    href: `/${lang}/my-spaces`,
+                  },
+                ]}
+              />
+            </MenuTop>
+            <NextSSRPlugin routerConfig={extractRouterConfig(fileRouter)} />
+            <div className="mb-auto pb-4">
+              <div className="pt-9 h-full flex justify-normal">
+                <div className="w-full h-full">{children}</div>
+              </div>
+            </div>
+            <Footer />
+          </EvmProvider>
+        </ThemeProvider>
+      </AuthProvider>
+      {shouldInjectToolbar && <VercelToolbar />}
+    </Html>
   );
 }
